pyproject.toml
[project]
name = "rag-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Kittipat Jusingha",email = "6531304621@student.chula.ac.th"}
]
readme = "README.md"
requires-python = ">=3.9,<3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "pymongo (>=4.13.0,<5.0.0)",
    "pinecone[grpc] (>=7.0.0,<8.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "google-genai (>=1.16.1,<2.0.0)",
    "pyreadline3 (>=3.5.4,<4.0.0)",
]

[tool.poetry]
packages = [
  { include = "streamlit", from = "src" },
  { include = "fastapi", from = "src" },
  { include = "model", from = "src" },
  { include = "pinecone", from = "src" },
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
